removed old rendering system
